+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.2.0
 
-- Fixed scrolling issue on 'plaintext' page.
+- Added "scroll to bottom" button to file content page.
+- Fixed scrolling issue on file content page.
 
 ## 0.1.0